refactor: placed discussion and notification constant on common place
