+++ src/courseware/course/sidebar/sidebars/discussions/DiscussionsTrigger.jsx
@@ -10,9 +10,10 @@ import { getCourseDiscussionTopics } from '../../../../data/thunks';
 import SidebarTriggerBase from '../../common/TriggerBase';
 import SidebarContext from '../../SidebarContext';
 import messages from './messages';
+import { WIDGETS } from '../../../../../generic/hooks';
 
 ensureConfig(['DISCUSSIONS_MFE_BASE_URL']);
-export const ID = 'DISCUSSIONS';
+export const ID = WIDGETS.DISCUSSIONS;
 
 const DiscussionsTrigger = ({
   intl,
